Remove duplicated spring-boot main file
