--- .vscode/launch.json
@@ -1,34 +1,0 @@
-{
-  "configurations": [
-    {
-      "type": "node",
-      "request": "attach",
-      "name": "Attach by Process ID",
-      "processId": "${command:PickProcess}"
-    },
-    {
-      "type": "chrome",
-      "request": "launch",
-      "name": "EmberJS",
-      "port": 9222,
-      "runtimeArgs": [
-        "--remote-debugging-port=9222"
-      ],
-      "url": "http://localhost:4200",
-      "sourceMapPathOverrides": {
-        "ember-java/*": "${workspaceFolder}/app/*"
-      }
-    },
-    {
-      "type": "java",
-      "name": "CodeLens (Launch) - Application",
-      "request": "launch",
-      "cwd": "${workspaceFolder}",
-      "console": "internalConsole",
-      "stopOnEntry": false,
-      "mainClass": "ember.sample.Application",
-      "args": "",
-      "projectName": "ember.sample"
-    }
-  ]
-}
